Split code from launchii.py into core.py
