+++ launchii/launchii.py
@@ -6,74 +6,14 @@ Usage:
 """
 import sys
 import platform
-import importlib
 import pathlib
-import json
-from typing import Any, List, Tuple
 
 import pinject
 import appdirs
 
-from launchii.api import Action, Item, Searcher, Solution
 import launchii.cli
 import launchii.gui
-
-_default_plugins = ["launchiicontrib.appsearch", "launchiicontrib.openaction"]
-
-
-class BasicSolution:
-    def __init__(self, item: Item, action: Action) -> None:
-        self.item = item
-        self.action = action
-
-    def describe(self) -> str:
-        return str(self.item.name)
-
-    def execute(self) -> Any:
-        return self.action.do(self.item)
-
-
-class BasicLaunchii:
-    def __init__(self, seacher: Searcher, action: Action) -> None:
-        self.searcher = seacher
-        self.action = action
-
-    def search(self, search_term: str) -> List[Solution]:
-        return list(
-            map(
-                lambda i: BasicSolution(i, self.action),
-                self.searcher.search(search_term),
-            )
-        )
-
-
-def load_plugin_file(config_dir: pathlib.Path, default) -> List[str]:
-    try:
-        with open(config_dir / "plugins.json") as f:
-            return default
-    except FileNotFoundError as err:
-        config_dir.mkdir(parents=True, exist_ok=True)
-        with open(config_dir / "plugins.json", "w") as f:
-            json.dump(default, f)
-        return default
-
-
-def instantiate_plugins(
-    packages: List[str],
-    instantiator,
-) -> Tuple[List[Searcher], List[Action]]:
-
-    searchers: List[Searcher] = []
-    actions: List[Action] = []
-
-    for package in packages:
-        actual_module = importlib.import_module(package)
-        index_class = getattr(actual_module, "Index")
-        index = instantiator.provide(index_class)
-        searchers.extend(map(instantiator.provide, index.searchers()))
-        actions.extend(map(instantiator.provide, index.actions()))
-
-    return (searchers, actions)
+from launchii.core import PluginLaunchii
 
 
 def main(cli, gui, print, args, launchii):
@@ -86,7 +26,7 @@ def main(cli, gui, print, args, launchii):
 
 
 def run():
-    class BaseProviderSpec(pinject.BindingSpec):
+    class BoostrapProviderSpec(pinject.BindingSpec):
         def provide_app_dirs(self):
             return appdirs.AppDirs("launchii")
 
@@ -96,13 +36,13 @@ def run():
         def provide_system(self):
             return platform.system()
 
-    instantiator = pinject.new_object_graph(binding_specs=[BaseProviderSpec()])
+        def provide_bootstrap_provider_spec(self):
+            return self
 
-    dirs = appdirs.AppDirs("launchii")
-    plugin_list = load_plugin_file(pathlib.Path(dirs.user_config_dir), _default_plugins)
+    instantiator = pinject.new_object_graph(
+        modules=None, binding_specs=[BoostrapProviderSpec()]
+    )
 
-    (searchers, actions) = instantiate_plugins(plugin_list, instantiator)
-
-    launchiiApp = BasicLaunchii(searchers[0], actions[0])
+    launchiiApp = instantiator.provide(PluginLaunchii)
 
     main(launchii.cli.Cli, launchii.gui.Gui, print, sys.argv, launchiiApp)
